Init Refactoring
