+++ queries/02.SEQUENCE & FUNCTION
@@ -1,0 +1,14 @@
+CREATE SEQUENCE td_num
+  START WITH 1
+  MAXVALUE 999999
+  INCREMENT BY 1
+  NOCYCLE
+  CACHE 20;
+  
+CREATE OR REPLACE FUNCTION td_code
+RETURN CHAR
+IS
+BEGIN
+    RETURN concat('td_',LPAD(n_num.nextval,5,0));
+END;
+/
