--- src/com/jspstudy/command/tCompleteCommand.java
@@ -1,19 +1,0 @@
-package com.jspstudy.command;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import com.jspstudy.dao.tDao;
-
-public class tCompleteCommand implements tCommand {
-
-	@Override
-	public void execute(HttpServletRequest request, HttpServletResponse response) {
-		String tId = request.getParameter("tId");
-		
-		tDao dao = new tDao();
-		
-		dao.complete(tId);
-	}
-
-}
