+++ src/controller/Controller.java
@@ -1,0 +1,10 @@
+package controller;
+
+public interface Controller {
+	
+	public void execute();
+	
+	public String moveUrl();
+	
+	public boolean isFoward();
+}
