+++ src/legacy/com/jspstudy/command/tAddCommand.java
@@ -1,0 +1,20 @@
+package legacy.com.jspstudy.command;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import legacy.com.jspstudy.dao.tDao;
+
+public class tAddCommand implements tCommand {
+
+	@Override
+	public void execute(HttpServletRequest request, HttpServletResponse response) {
+		
+		String tContent = request.getParameter("todo");
+		
+		tDao dao = new tDao();
+		
+		dao.add(tContent);
+	}
+
+}
