+++ src/legacy/com/jspstudy/command/tCommand.java
@@ -1,0 +1,10 @@
+package legacy.com.jspstudy.command;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public interface tCommand {
+	
+	void execute(HttpServletRequest request, HttpServletResponse response);
+
+}
