+++ src/legacy/com/jspstudy/dao/tDao.java
@@ -1,0 +1,156 @@
+package legacy.com.jspstudy.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.Timestamp;
+import java.util.ArrayList;
+
+import javax.naming.Context;
+import javax.naming.InitialContext;
+import javax.sql.DataSource;
+
+import legacy.com.jspstudy.dto.tDto;
+
+public class tDao {
+	
+	DataSource dataSource;
+	
+	public tDao() {
+		
+		try {
+			Context context = new InitialContext();
+			dataSource = (DataSource)context.lookup("java:comp/env/jdbc/Oracle11g");
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		
+	}
+	
+	public void add(String tContent) {
+		
+		Connection connection = null;
+		PreparedStatement preparedStatement = null;
+		
+		try {
+			connection = dataSource.getConnection();
+			String query = ""
+					+ "insert into todo(tId, tContent, tComplete) "
+					+ "values (todo_seq.nextval, ?, 0)";
+			
+			preparedStatement = connection.prepareStatement(query);
+			preparedStatement.setString(1, tContent);
+			int rn = preparedStatement.executeUpdate();
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				if (preparedStatement != null) preparedStatement.close();
+				if (connection != null) connection.close();
+			} catch (Exception e2) {
+				e2.printStackTrace();
+			}
+		}
+	}
+	
+	public ArrayList<tDto> list() {
+		
+		ArrayList<tDto> dtos = new ArrayList<tDto>();
+		
+		Connection connection = null;
+		PreparedStatement preparedStatement = null;
+		ResultSet resultSet = null;
+		
+		try {
+			connection = dataSource.getConnection();
+			
+			String query = ""
+					+ "select tId, tContent, tComplete, sDate "
+					+ "from todo "
+					+ "order by tId desc";
+			preparedStatement = connection.prepareStatement(query);
+			resultSet = preparedStatement.executeQuery();
+			
+			while(resultSet.next()) {
+				int tId = resultSet.getInt("tId");
+				String tContent = resultSet.getString("tContent");
+				int tComplete = resultSet.getInt("tComplete");
+				Timestamp sDate = resultSet.getTimestamp("sDate");
+				
+				tDto dto = new tDto(tId, tContent, tComplete, sDate);
+				dtos.add(dto);
+			}
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				if(resultSet != null) resultSet.close();
+				if(preparedStatement != null) preparedStatement.close();
+				if(connection != null) connection.close();
+			} catch (Exception e2) {
+				e2.printStackTrace();
+			}
+		}
+		
+		return dtos;
+	}
+	
+	public void delete(String tId) {
+		
+		Connection connection = null;
+		PreparedStatement preparedStatement = null;
+		
+		try {
+			connection = dataSource.getConnection();
+			String query = ""
+					+ "delete from todo "
+					+ "where tId = ?";
+			
+			preparedStatement = connection.prepareStatement(query);
+			preparedStatement.setInt(1, Integer.parseInt(tId));
+			int rn = preparedStatement.executeUpdate();
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				if(preparedStatement != null) preparedStatement.close();
+				if(connection != null) preparedStatement.close();
+			} catch (Exception e2) {
+				e2.printStackTrace();
+			}
+		}
+	}
+	
+	public void complete(String tId) {
+		
+		Connection connection = null;
+		PreparedStatement preparedStatement = null;
+		
+		try {
+			connection = dataSource.getConnection();
+			String query = ""
+					+ "update todo "
+					+ "set tComplete = 1 "
+					+ "where tId = ?";
+			
+			preparedStatement = connection.prepareStatement(query);
+			preparedStatement.setInt(1, Integer.parseInt(tId));
+			int rn = preparedStatement.executeUpdate();
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				if(preparedStatement != null) preparedStatement.close();
+				if(connection != null) connection.close();
+			} catch (Exception e2) {
+				e2.printStackTrace();
+			}
+		}
+	}
+
+	
+}
