+++ src/service/ToDoService.java
@@ -1,0 +1,25 @@
+package service;
+
+public class ToDoService {
+	
+	public boolean addToDo(String content) { 
+		boolean flag = false;
+		
+		
+		return flag;
+	}
+
+	public boolean finishToDo(String td_num) {
+		boolean flag = false;
+		
+		
+		return flag;
+	}
+	
+	public boolean removeToDo(String td_num) {
+		boolean flag = false;
+		
+		
+		return flag;
+	}
+}
